Add jsdoc, remove old i18n code
