+++ packages/i18n/test/server.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "vitest";
-import * as i18n from "../src/new-i18n.js";
+import * as i18n from "../src/index.js";
 import { en_dict } from "./setup.js";
 
 describe("template", () => {
