+++ packages/i18n/test/setup.ts
@@ -1,4 +1,4 @@
-import * as i18n from "../src/new-i18n.js";
+import * as i18n from "../src/index.js";
 
 export class MyClass {
   constructor(public name: string) {}
